refactor: chunk out course-parsing logic into smaller units
Maintaining and testing is easier if the logic is separated by object.
